# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "东南大学成贤学院学生成绩录入及学生进行查询系统的标准API接口文档",
        "title": "成绩录入及查询系统标准接口文档",
        "termsOfService": "暂缺",
        "contact": {
            "name": "黄浩轩",
            "url": "https://github.com/huanghaoxuan",
            "email": "huanghaoxuan1998@outlook.com"
        },
        "license": {
          "name": "GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007",
          "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:5201",
    "basePath": "/api",
    "paths": {
...
      "/analysis/teachingclass/{name}": {
        "get": {
          "description": "教师分析任课班课程情况",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "首页数据源"
          ],
          "summary": "教师分析任课班课程情况",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "教师姓名",
              "name": "name",
              "in": "path",
              "required": true
            }
          ]
        }
      },
...
      "/teachingClassInformation/delete/{id}": {
        "get": {
          "description": "删除一条教学班信息",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "教学班基本信息"
          ],
          "summary": "删除一条教学班信息",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "主键",
              "name": "id",
              "in": "path",
              "required": true
            }
          ]
        }
      },
      "/teachingClassInformation/insert": {
        "post": {
          "description": "增加教学班信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "教学班基本信息"
          ],
          "summary": "增加教学班信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "课程名称",
              "name": "courseName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "课程id",
              "name": "courseId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师名字",
              "name": "courseTeacherName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师id",
              "name": "courseTeacherId",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
      "/teachingClassInformation/selectByPage": {
        "post": {
          "description": "分页查询教学班信息，如果查询第一页，返回总条数，条件非必需",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "教学班基本信息"
          ],
          "summary": "分页查询教学班信息",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "查询页码",
              "name": "pageNum",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "每页条数",
              "name": "pageSize",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "课程名称",
              "name": "courseName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "课程id",
              "name": "courseId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师名字",
              "name": "courseTeacherName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师id",
              "name": "courseTeacherId",
              "in": "formData"
            }
          ]
        }
      },
      "/teachingClassInformation/selectCrossSemester": {
        "post": {
          "description": "分页查询教学班是否存在跨学期内容",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "教学班信息"
          ],
          "summary": "分页查询教学班是否存在跨学期内容",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "查询页码",
              "name": "pageNum",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "每页条数",
              "name": "pageSize",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "学生学号",
              "name": "studentId",
              "in": "formData",
              "required": true
            },
            {
              "type": "string",
              "description": "姓名",
              "name": "name",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在年级",
              "name": "grade",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在学院或部门",
              "name": "department",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在专业",
              "name": "professional",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "所在班级",
              "name": "class",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "课程名称",
              "name": "courseName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "课程id",
              "name": "courseId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师名字",
              "name": "courseTeacherName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师id",
              "name": "courseTeacherId",
              "in": "formData"
            }
          ]
        }
      },
      "/teachingClassInformation/update": {
        "post": {
          "description": "更新教学班信息记录",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "教学班基本信息"
          ],
          "summary": "更新教学班信息记录",
          "parameters": [
            {
              "type": "string",
              "description": "Token",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "description": "课程名称",
              "name": "courseName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "课程id",
              "name": "courseId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "教学班号",
              "name": "teachingClassId",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师名字",
              "name": "courseTeacherName",
              "in": "formData"
            },
            {
              "type": "string",
              "description": "任课老师id",
              "name": "courseTeacherId",
              "in": "formData"
            }
          ],
          "responses": {
            "200": {
              "description": "{\"status\": 1}",
              "schema": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
}